+++ src/windows/SQLite3-Win-RT/SQLite3/Statement.h
@@ -30,6 +30,7 @@ namespace SQLite3
     int BindNull(int index);
 
   private:
+    Database^ db;
     sqlite3_stmt* statement;
   };
 }